Replaced Hero.js hardcoded Circles component to object literal mapping
What title says.

But we still need to find a way to replace the icons within the return component.

I tried <${icons.icon} to see if that returns an identifier, but it didn't.

We should discuss if we want to move the hardcoded object literals to our backend or database as well.

Worst comes worst, we can always use if/else statement within Circles component to identy the icons